Update workspace configuration and add utils module for transitions
