+++ Cargo.toml
@@ -69,5 +69,5 @@ panic = "abort"   # Higher performance by disabling panic handlers.
 strip = true      # Ensures debug symbols are removed.
 
 [workspace]
-members = ["crates/route_transitions", "."]
+members = ["crates/route_transitions", ".", "router_test"]
 resolver = "2"
